+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.1] - 2020-02-23
+
 ### Added
 
 - `--more` option for worktime query
@@ -142,7 +144,8 @@ First release :tada:
 
 - Linux binaries
 
-[unreleased]: https://github.com/unfog-io/unfog-cli/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/unfog-io/unfog-cli/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/unfog-io/unfog-cli/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/unfog-io/unfog-cli/compare/v0.3.3...v0.4.0
 [0.3.3]: https://github.com/unfog-io/unfog-cli/compare/v0.3.2...v0.3.3
 [0.3.2]: https://github.com/unfog-io/unfog-cli/compare/v0.3.1...v0.3.2